Create user, freeboard service
