+++ src/main/java/com/springboot/intelllij/domain/LensEntity.java
@@ -5,7 +5,6 @@ import com.vladmihalcea.hibernate.type.json.JsonBinaryType;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.TypeDef;
-import springfox.documentation.spring.web.json.Json;
 
 import javax.persistence.*;
