+++ src/main/java/com/springboot/intelllij/services/LensInfoService.java
@@ -4,7 +4,6 @@ import com.springboot.intelllij.domain.LensEntity;
 import com.springboot.intelllij.exceptions.LensNotFoundException;
 import com.springboot.intelllij.repository.LensRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
